// packages/study-echarts/src/echarts/util/types.ts
export type HorizontalAlign = 'left' | 'center' | 'right';
export type VerticalAlign = 'top' | 'middle' | 'bottom';

export type ColorString = string;
// ...
export type ZRRectLike = RectLike;

export type ZRStyleProps = PathStyleProps | ImageStyleProps | TSpanStyleProps | TextStyleProps;
// ...
export type OptionDataValue = string | number | Date;
// ...
export interface BoxLayoutOptionMixin {
  width?: number | string;
  height?: number | string;
  top?: number | string;
  right?: number | string;
  bottom?: number | string;
  left?: number | string;
}
// ...
export type AnimationDelayCallbackParam = {
  count: number
  index: number
};
export type AnimationDurationCallback = (idx: number) => number;
export type AnimationDelayCallback = (idx: number, params?: AnimationDelayCallbackParam) => number;

export interface AnimationOption {
  duration?: number
  easing?: AnimationEasing
  delay?: number
  // additive?: boolean
}

/**
* Mixin of option set to control the animation of series.
*/
export interface AnimationOptionMixin {
  /**
   * If enable animation
   */
  animation?: boolean
  /**
   * Disable animation when the number of elements exceeds the threshold
   */
  animationThreshold?: number
  // For init animation
  /**
   * Duration of initialize animation.
   * Can be a callback to specify duration of each element
   */
  animationDuration?: number | AnimationDurationCallback
  /**
   * Easing of initialize animation
   */
  animationEasing?: AnimationEasing
  /**
   * Delay of initialize animation
   * Can be a callback to specify duration of each element
   */
  animationDelay?: number | AnimationDelayCallback
  // For update animation
  /**
   * Delay of data update animation.
   * Can be a callback to specify duration of each element
   */
  animationDurationUpdate?: number | AnimationDurationCallback
  /**
   * Easing of data update animation.
   */
  animationEasingUpdate?: AnimationEasing
  /**
   * Delay of data update animation.
   * Can be a callback to specify duration of each element
   */
  animationDelayUpdate?: number | AnimationDelayCallback
}
// ...
export interface TextCommonOption extends ShadowOptionMixin {
  color?: string
  fontStyle?: ZRFontStyle
  fontWeight?: ZRFontWeight
  fontFamily?: string
  fontSize?: number | string
  align?: HorizontalAlign
  verticalAlign?: VerticalAlign
  // @deprecated
  baseline?: VerticalAlign

  opacity?: number

  lineHeight?: number
  backgroundColor?: ColorString | {
    image: ImageLike | string
  }
  borderColor?: string
  borderWidth?: number
  borderType?: ZRLineType
  borderDashOffset?: number
  borderRadius?: number | number[]
  padding?: number | number[]

  width?: number | string// Percent
  height?: number
  textBorderColor?: string
  textBorderWidth?: number
  textBorderType?: ZRLineType
  textBorderDashOffset?: number

  textShadowBlur?: number
  textShadowColor?: string
  textShadowOffsetX?: number
  textShadowOffsetY?: number

  tag?: string
}

export interface LabelFormatterCallback<T = CallbackDataParams> {
  (params: T): string
}
/**
* LabelOption is an option set to control the style of labels.
* Include color, background, shadow, truncate, rotation, distance, etc..
*/
export interface LabelOption extends TextCommonOption {
  /**
   * If show label
   */
  show?: boolean
  // TODO: TYPE More specified 'inside', 'insideTop'....
  // x, y can be both percent string or number px.
  position?: ElementTextConfig['position']
  distance?: number
  rotate?: number
  offset?: number[]

  /**
   * Min margin between labels. Used when label has layout.
   */
  // It's minMargin instead of margin is for not breaking the previous code using margin.
  minMargin?: number

  overflow?: TextStyleProps['overflow']
  silent?: boolean
  precision?: number | 'auto'
  valueAnimation?: boolean

  // TODO: TYPE not all label support formatter
  // formatter?: string | ((params: CallbackDataParams) => string)

  rich?: Dictionary<TextCommonOption>
}
// ...
export interface TooltipFormatterCallback<T> {
  /**
   * For sync callback
   * params will be an array on axis trigger.
   */
  (params: T, asyncTicket: string): string | HTMLElement | HTMLElement[]
  /**
   * For async callback.
   * Returned html string will be a placeholder when callback is not invoked.
   */
  (
    params: T, asyncTicket: string,
    callback: (cbTicket: string, htmlOrDomNodes: string | HTMLElement | HTMLElement[]) => void
  ): string | HTMLElement | HTMLElement[]
}

type TooltipBuiltinPosition = 'inside' | 'top' | 'left' | 'right' | 'bottom';
type TooltipBoxLayoutOption = Pick<
  BoxLayoutOptionMixin, 'top' | 'left' | 'right' | 'bottom'
>;

export type TooltipPositionCallbackParams = CallbackDataParams | CallbackDataParams[];

/**
* Position relative to the hoverred element. Only available when trigger is item.
*/
export interface TooltipPositionCallback {
  (
    point: [number, number],
    /**
     * params will be an array on axis trigger.
     */
    params: TooltipPositionCallbackParams,
    /**
     * Will be HTMLDivElement when renderMode is html
     * Otherwise it's graphic.Text
     */
    el: HTMLDivElement | ZRText | null,
    /**
     * Rect of hover elements. Will be null if not hovered
     */
    rect: RectLike | null,
    size: {
      /**
       * Size of popup content
       */
      contentSize: [number, number]
      /**
       * Size of the chart view
       */
      viewSize: [number, number]
    }
  ): Array<number | string> | TooltipBuiltinPosition | TooltipBoxLayoutOption
}
/**
* Common tooltip option
* Can be configured on series, graphic elements
*/
export interface CommonTooltipOption<FormatterParams> {

  show?: boolean

  /**
   * When to trigger
   */
  triggerOn?: 'mousemove' | 'click' | 'none' | 'mousemove|click'
  /**
   * Whether to not hide popup content automatically
   */
  alwaysShowContent?: boolean

  formatter?: string | TooltipFormatterCallback<FormatterParams>

  /**
   * Formatter of value.
   *
   * Will be ignored if tooltip.formatter is specified.
   */
  valueFormatter?: (value: OptionDataValue | OptionDataValue[]) => string
  /**
   * Absolution pixel [x, y] array. Or relative percent string [x, y] array.
   * If trigger is 'item'. position can be set to 'inside' / 'top' / 'left' / 'right' / 'bottom',
   * which is relative to the hovered element.
   *
   * Support to be a callback
   */
  position?: (number | string)[] | TooltipBuiltinPosition | TooltipPositionCallback | TooltipBoxLayoutOption

  confine?: boolean

  /**
   * Consider triggered from axisPointer handle, verticalAlign should be 'middle'
   */
  align?: HorizontalAlign

  verticalAlign?: VerticalAlign
  /**
   * Delay of show. milesecond.
   */
  showDelay?: number

  /**
   * Delay of hide. milesecond.
   */
  hideDelay?: number

  transitionDuration?: number
  /**
   * Whether mouse is allowed to enter the floating layer of tooltip
   * If you need to interact in the tooltip like with links or buttons, it can be set as true.
   */
  enterable?: boolean

  backgroundColor?: ColorString
  borderColor?: ColorString
  borderRadius?: number
  borderWidth?: number
  shadowBlur?: number
  shadowColor?: string
  shadowOffsetX?: number
  shadowOffsetY?: number

  /**
   * Padding between tooltip content and tooltip border.
   */
  padding?: number | number[]

  /**
   * Available when renderMode is 'html'
   */
  extraCssText?: string

  textStyle?: Pick<LabelOption,
    'color' | 'fontStyle' | 'fontWeight' | 'fontFamily' | 'fontSize' |
    'lineHeight' | 'width' | 'height' | 'textBorderColor' | 'textBorderWidth' |
    'textShadowColor' | 'textShadowBlur' | 'textShadowOffsetX' | 'textShadowOffsetY'
    | 'align'> & {

      // Available when renderMode is html
      decoration?: string
    }
}

export type ComponentItemTooltipOption<T> = CommonTooltipOption<T> & {
  // Default content HTML.
  content?: string;
  formatterParams?: ComponentItemTooltipLabelFormatterParams;
};
export type ComponentItemTooltipLabelFormatterParams = {
  componentType: string
  name: string
  // properies key array like ['name']
  $vars: string[]
} & {
    // Other properties
    [key in string]: unknown
  };